+++ android/src/main/kotlin/com/gstory/flutter_unionad/FlutterUnionadViewPlugin.kt
@@ -3,7 +3,6 @@ package com.gstory.flutter_unionad
 import android.app.Activity
 import com.gstory.flutter_unionad.bannerad.BannerExpressAdViewFactory
 import com.gstory.flutter_unionad.drawfeedad.DrawFeedExpressAdViewFactory
-import com.gstory.flutter_unionad.interactionad.InteractionExpressAdViewFactory
 import com.gstory.flutter_unionad.nativead.NativeExpressAdFactory
 import com.gstory.flutter_unionad.splashad.SplashAdViewFactory
 import io.flutter.embedding.engine.plugins.FlutterPlugin
@@ -21,8 +20,6 @@ object FlutterUnionadViewPlugin {
         binding.platformViewRegistry.registerViewFactory(FlutterunionadViewConfig.bannerAdView, BannerExpressAdViewFactory(binding.binaryMessenger,activity))
         //注册个性化模板信息流广告
         binding.platformViewRegistry.registerViewFactory(FlutterunionadViewConfig.nativeAdView, NativeExpressAdFactory(binding.binaryMessenger,activity))
-        //注册个性化模板插屏广告
-        binding.platformViewRegistry.registerViewFactory(FlutterunionadViewConfig.interactionAdVie, InteractionExpressAdViewFactory(binding.binaryMessenger,activity))
         //注册个性化模板draw视频广告
         binding.platformViewRegistry.registerViewFactory(FlutterunionadViewConfig.drawFeedAdView, DrawFeedExpressAdViewFactory(binding.binaryMessenger,activity))
     }