+++ ios/Classes/nativead/NativeAdViewFactory.swift
@@ -15,11 +15,11 @@ public class NativeAdViewFactory: NSObject,FlutterPlatformViewFactory {
         self.messenger = messenger
         super.init()
     }
-    
+
     public func create(withFrame frame: CGRect, viewIdentifier viewId: Int64, arguments args: Any?) -> FlutterPlatformView {
         return NativeAdView(frame,  binaryMessenger: self.messenger,id: viewId, params:args)
     }
-    
+
     public func createArgsCodec() -> FlutterMessageCodec & NSObjectProtocol {
         return FlutterStandardMessageCodec.sharedInstance()
     }