+++ lib/nativead/NativeAdView.dart
@@ -10,9 +10,7 @@ class NativeAdView extends StatefulWidget {
   final bool supportDeepLink;
   final double expressViewWidth;
   final double expressViewHeight;
-  final int expressNum;
   final int downloadType;
-  final bool isUserInteractionEnabled;
   final int? adLoadType;
   final FlutterUnionadNativeCallBack? callBack;
 
@@ -24,9 +22,7 @@ class NativeAdView extends StatefulWidget {
       required this.supportDeepLink,
       required this.expressViewWidth,
       required this.expressViewHeight,
-      required this.expressNum,
       required this.downloadType,
-      required this.isUserInteractionEnabled,
       required this.adLoadType,
       this.callBack})
       : super(key: key);
@@ -72,7 +68,6 @@ class _NativeAdViewState extends State<NativeAdView> {
             "supportDeepLink": widget.supportDeepLink,
             "expressViewWidth": widget.expressViewWidth,
             "expressViewHeight": widget.expressViewHeight,
-            "expressNum": widget.expressNum,
             "downloadType": widget.downloadType,
             "adLoadType": widget.adLoadType,
           },
@@ -81,8 +76,6 @@ class _NativeAdViewState extends State<NativeAdView> {
         ),
       );
     } else if (defaultTargetPlatform == TargetPlatform.iOS) {
-      sendMsgToNative('isUserInteractionEnabled',
-          body: widget.isUserInteractionEnabled);
       return Container(
         width: _width,
         height: _height == 0 ? 0.5 : _height, //高为0的时候原生不会加载 默认设为0.5
@@ -94,7 +87,6 @@ class _NativeAdViewState extends State<NativeAdView> {
             "supportDeepLink": widget.supportDeepLink,
             "expressViewWidth": widget.expressViewWidth,
             "expressViewHeight": widget.expressViewHeight,
-            "expressNum": widget.expressNum,
             "downloadType": widget.downloadType,
           },
           onPlatformViewCreated: _registerChannel,
@@ -110,17 +102,6 @@ class _NativeAdViewState extends State<NativeAdView> {
   void _registerChannel(int id) {
     _channel = MethodChannel("${_viewType}_$id");
     _channel?.setMethodCallHandler(_platformCallHandler);
-  }
-
-  Future sendMsgToNative(String name, {Object? body}) async {
-    try {
-      var result = await _channel?.invokeMethod(name, body);
-      return result;
-    } on PlatformException catch (e) {
-      Future fut = Future.error(e.toString());
-      FlutterError.reportError(FlutterErrorDetails(exception: e));
-      return fut;
-    }
   }
 
   //监听原生view传值